+++ Demos/02-fieldcustomizer/.yo-rc.json
@@ -2,11 +2,10 @@
   "@microsoft/generator-sharepoint": {
     "isCreatingSolution": true,
     "environment": "spo",
-    "version": "1.7.0",
+    "version": "1.7.1",
     "libraryName": "sp-fx-field-customizer",
-    "libraryId": "b6b8743e-cee7-4c8f-ab63-e479e80d4567",
+    "libraryId": "c5162338-7894-48af-abef-c93249bad7b4",
     "packageManager": "npm",
-    "isDomainIsolated": false,
     "componentType": "extension",
     "extensionType": "FieldCustomizer"
   }
