+++ Demos/02-fieldcustomizer/package.json
@@ -11,17 +11,17 @@
     "test": "gulp test"
   },
   "dependencies": {
-    "@microsoft/sp-core-library": "1.7.0",
-    "@microsoft/decorators": "1.7.0",
+    "@microsoft/sp-core-library": "1.7.1",
+    "@microsoft/decorators": "1.7.1",
     "@types/webpack-env": "1.13.1",
     "@types/es6-promise": "0.0.33",
-    "@microsoft/sp-listview-extensibility": "1.7.0"
+    "@microsoft/sp-listview-extensibility": "1.7.1"
   },
   "devDependencies": {
-    "@microsoft/sp-build-web": "1.7.0",
-    "@microsoft/sp-tslint-rules": "1.7.0",
-    "@microsoft/sp-module-interfaces": "1.7.0",
-    "@microsoft/sp-webpart-workbench": "1.7.0",
+    "@microsoft/sp-build-web": "1.7.1",
+    "@microsoft/sp-tslint-rules": "1.7.1",
+    "@microsoft/sp-module-interfaces": "1.7.1",
+    "@microsoft/sp-webpart-workbench": "1.7.1",
     "gulp": "~3.9.1",
     "@types/chai": "3.4.34",
     "@types/mocha": "2.2.38",
