+++ Demos/02-fieldcustomizer/sharepoint/assets/elements.xml
@@ -1,13 +1,13 @@
 <?xml version="1.0" encoding="utf-8"?>
 <Elements xmlns="http://schemas.microsoft.com/sharepoint/">
-    <Field ID="{0ebe7606-20f4-49be-bcdc-2dfc65348b46}"
+    <Field ID="{c5db0a97-b8fa-4c88-b14f-a0c20684c91c}"
             Name="SPFxPercentage"
-            DisplayName="Percentage Complete"
+            DisplayName="Percentage"
             Type="Number"
             Min="0"
             Required="FALSE"
             Group="SPFx Columns"
-            ClientSideComponentId="fcedd96a-1c34-4ac8-9ad8-5aaf4cb1e993"
+            ClientSideComponentId="6a1b8997-00d5-4bc7-a472-41d6ac27cd83"
             ClientSideComponentProperties="{&quot;greenMinLimit&quot;:&quot;85&quot;,&quot;yellowMinLimit&quot;:&quot;70&quot;}">
     </Field>
 </Elements>
