+++ Demos/02-fieldcustomizer/src/extensions/helloFieldCustomizer/loc/en-us.js
@@ -1,0 +1,5 @@
+define([], function() {
+  return {
+    "Title": "HelloFieldCustomizerFieldCustomizer"
+  }
+});
