+++ Demos/02-fieldcustomizer/src/extensions/helloFieldCustomizer/loc/myStrings.d.ts
@@ -1,0 +1,8 @@
+declare interface IHelloFieldCustomizerFieldCustomizerStrings {
+  Title: string;
+}
+
+declare module 'HelloFieldCustomizerFieldCustomizerStrings' {
+  const strings: IHelloFieldCustomizerFieldCustomizerStrings;
+  export = strings;
+}
