--- Demos/02-fieldcustomizer/src/extensions/percentComplete/PercentCompleteFieldCustomizer.manifest.json
@@ -1,17 +1,0 @@
-{
-  "$schema": "https://developer.microsoft.com/json-schemas/spfx/client-side-extension-manifest.schema.json",
-
-  "id": "fcedd96a-1c34-4ac8-9ad8-5aaf4cb1e993",
-  "alias": "PercentCompleteFieldCustomizer",
-  "componentType": "Extension",
-  "extensionType": "FieldCustomizer",
-
-  // The "*" signifies that the version should be taken from the package.json
-  "version": "*",
-  "manifestVersion": 2,
-
-  // If true, the component can only be installed on sites where Custom Script is allowed.
-  // Components that allow authors to embed arbitrary script code should set this to true.
-  // https://support.office.com/en-us/article/Turn-scripting-capabilities-on-or-off-1f2c515f-5d7e-448a-9fd7-835da935584f
-  "requiresCustomScript": false
-}
