--- Demos/02-fieldcustomizer/src/extensions/percentComplete/PercentCompleteFieldCustomizer.module.scss
@@ -1,9 +1,0 @@
-.PercentComplete {
-  .cell {
-    display: 'inline-block';
-  }
-  .filledBackground {
-    background-color: #cccccc;
-    width: 100px;
-  }
-}
