--- Demos/02-fieldcustomizer/src/extensions/percentComplete/loc/en-us.js
@@ -1,5 +1,0 @@
-define([], function() {
-  return {
-    "Title": "PercentCompleteFieldCustomizer"
-  }
-});
