--- Demos/02-fieldcustomizer/src/extensions/percentComplete/loc/myStrings.d.ts
@@ -1,8 +1,0 @@
-declare interface IPercentCompleteFieldCustomizerStrings {
-  Title: string;
-}
-
-declare module 'PercentCompleteFieldCustomizerStrings' {
-  const strings: IPercentCompleteFieldCustomizerStrings;
-  export = strings;
-}
